pyproject.toml
[project]
name = "descriptor-api"
version = "0.1.0"
description = "An API made for image captioning"
authors = [
    {name = "NotisFobbidden",email = "fobbidennotis@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "pyrogram",
  "aiogram",
  "easyocr",
  "pillow",
  "opencv-python",
  "transformers",
  "torch",
  "flask[async]"
]
package-mode = "false"


[tool.pyright]
venvPath = "."
venv = ".venv"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
